backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.1.0,<5.0.0
python-multipart>=0.0.9
httpx>=0.26.0
pillow>=10.3.0
aiofiles>=23.2.1
google-auth>=2.20
google-genai>=1.0
